@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xmlns="http://maven.apache.org/POM/4.0.0"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
@@ -9,12 +9,12 @@
     <version>0.0.1-SNAPSHOT</version>
 
     <properties>
-       <!-- <matsim.version>14.0-PR1525</matsim.version> -->
-        <matsim.version>14.0-SNAPSHOT</matsim.version>
+        <!-- <matsim.version>14.0-PR1525</matsim.version> -->
+        <matsim.version>15.0-SNAPSHOT</matsim.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <maven.compiler.source>11</maven.compiler.source>
-        <maven.compiler.target>11</maven.compiler.target>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
     </properties>
 
     <dependencies>
@@ -60,9 +60,9 @@
         </dependency>
 
         <dependency>
-        <groupId>com.github.matsim-scenarios</groupId>
-        <artifactId>matsim-berlin</artifactId>
-        <version>5.6.0</version>
+            <groupId>com.github.matsim-scenarios</groupId>
+            <artifactId>matsim-berlin</artifactId>
+            <version>5.6.0</version>
             <exclusions>
                 <exclusion>
                     <groupId>*</groupId>
@@ -70,7 +70,6 @@
                 </exclusion>
             </exclusions>
         </dependency>
-
 
 
         <dependency>
@@ -132,7 +131,8 @@
                             <outputFile>${project.basedir}/${project.build.finalName}.jar</outputFile>
                             <transformers>
                                 <!-- This bit sets the main class for the executable jar as you otherwise would with the assembly plugin -->
-                                <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
+                                <transformer
+                                        implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
                                     <manifestEntries>
                                         <Main-Class>org.matsim.gui.MATSimGUI</Main-Class>
                                         <Specification-Vendor>org.matsim</Specification-Vendor>
@@ -141,7 +141,8 @@
                                     </manifestEntries>
                                 </transformer>
                                 <!-- This bit merges the various GeoTools META-INF/services files         -->
-                                <transformer implementation="org.apache.maven.plugins.shade.resource.ServicesResourceTransformer"/>
+                                <transformer
+                                        implementation="org.apache.maven.plugins.shade.resource.ServicesResourceTransformer"/>
                             </transformers>
                             <filters>
                                 <filter>
